split constants files; split rarity & metadata into different components; add defaults values for network.js options
