+++ constants/rarity.js
@@ -1,0 +1,9 @@
+const RARITY = {
+  none: 0,
+  JaccardDistances: 1, // most accurate / recommended
+  TraitRarity: 2,
+  StatisticalRarity: 3,
+  TraitAndStatisticalRarity: 4,
+};
+
+module.exports = { RARITY };
