+++ utils/common.js
@@ -1,38 +1,28 @@
 const basePath = process.cwd();
 const fs = require("fs");
 const { network } = require(`${basePath}/src/config.js`);
-const { metadataTypes } = require(`${basePath}/constants/network.js`);
+const { METADATA } = require(`${basePath}/constants/metadata.js`);
 
 // get metadata of all generated NFTs/items
 const getMetadataItems = () => {
-  if (network.metadataType == metadataTypes.basic) {
+  if (network.metadataType === METADATA.basic) {
     // get metadata from the _metadata.json file
     return JSON.parse(
       fs.readFileSync(
-        `${basePath}/build/${network.jsonDirPrefix ?? ""}${
-          network.metadataFileName
-        }`
+        `${basePath}/build/${network.jsonDirPrefix}${network.metadataFileName}`
       )
     );
   } else {
     // get metadata from the individual metadata files
-    let rawData = [];
     const jsonFilePattern = /^\d+.(json)$/i;
-    const files = fs.readdirSync(
-      `${basePath}/build/${network.jsonDirPrefix ?? ""}`
-    );
-    files.forEach((file) => {
-      if (file.match(jsonFilePattern)) {
-        rawData.push(
-          JSON.parse(
-            fs.readFileSync(
-              `${basePath}/build/${network.jsonDirPrefix ?? ""}${file}`
-            )
-          )
-        );
-      }
-    });
-    return rawData;
+    const files = fs.readdirSync(`${basePath}/build/${network.jsonDirPrefix}`);
+    return files
+      .filter((file) => file.match(jsonFilePattern))
+      .map((file) =>
+        JSON.parse(
+          fs.readFileSync(`${basePath}/build/${network.jsonDirPrefix}${file}`)
+        )
+      );
   }
 };
