+++ utils/scripts/pixelate.js
@@ -4,7 +4,7 @@ const { createCanvas, loadImage } = require("canvas");
 const basePath = process.cwd();
 const { format, network, pixelFormat } = require(`${basePath}/src/config.js`);
 const buildDir = `${basePath}/build/pixel_images`;
-const inputDir = `${basePath}/build/${network.mediaDirPrefix ?? ""}`;
+const inputDir = `${basePath}/build/${network.mediaDirPrefix}`;
 const console = require("console");
 const canvas = createCanvas(format.width, format.height);
 const ctx = canvas.getContext("2d");
