+++ utils/scripts/preview_gif.js
@@ -2,7 +2,7 @@ const basePath = process.cwd();
 const fs = require("fs");
 const { createCanvas, loadImage } = require("canvas");
 const buildDir = `${basePath}/build`;
-const imageDir = `${buildDir}/${network.mediaDirPrefix ?? ""}`;
+const imageDir = `${buildDir}/${network.mediaDirPrefix}`;
 const { format, preview_gif } = require(`${basePath}/src/config.js`);
 const canvas = createCanvas(format.width, format.height);
 const ctx = canvas.getContext("2d");
