+++ utils/scripts/update_info.js
@@ -1,7 +1,7 @@
 const basePath = process.cwd();
 const { NETWORK } = require(`${basePath}/constants/network.js`);
 const fs = require("fs");
-const { metadataTypes } = require("../constants/network");
+const { METADATA } = require(`${basePath}/constants/metadata`);
 const { getMetadataItems } = require(`${basePath}/utils/common.js`);
 
 const {
@@ -19,36 +19,36 @@ let idx = network.startIdx;
 data.forEach((item) => {
   // general metadata
   item.description = description;
-  if (network.metadataType != metadataTypes.basic) {
+  if (network.metadataType != METADATA.basic) {
     item.name = `${namePrefix} #${idx++}`;
   } else {
     item.name = `${namePrefix} #${item.edition}`;
   }
 
   // custom metadata
-  if (network == NETWORK.eth) {
+  if (network === NETWORK.eth) {
     item.image = `${baseUri}/${item.edition}.png`;
   }
-  if (network == NETWORK.sol) {
+  if (network === NETWORK.sol) {
     item.creators = solanaMetadata.creators;
   }
 
   fs.writeFileSync(
-    `${basePath}/build/${network.jsonDirPrefix ?? ""}${item.edition}.json`,
+    `${basePath}/build/${network.jsonDirPrefix}${item.edition}.json`,
     JSON.stringify(item, null, 2)
   );
 });
 
-if (network.metadataType == metadataTypes.basic) {
+if (network.metadataType === METADATA.basic) {
   fs.writeFileSync(
-    `${basePath}/build/${network.jsonDirPrefix ?? ""}${
+    `${basePath}/build/${network.jsonDirPrefix}${
       network.metadataFileName
     }`,
     JSON.stringify(data, null, 2)
   );
 }
 
-if (network == NETWORK.sol) {
+if (network === NETWORK.sol) {
   console.log(`Updated description for images to ===> ${description}`);
   console.log(`Updated name prefix for images to ===> ${namePrefix}`);
   console.log(
